Rename repo to github-store from github-kv
charliesome preferred something more generic than kv and I was indifferent. Should be the last rename... :)
