+++ CONTRIBUTING.md
@@ -1,7 +1,7 @@
 ## Contributing
 
-[fork]: https://github.com/github/github-kv/fork
-[pr]: https://github.com/github/github-kv/compare
+[fork]: https://github.com/github/github-store/fork
+[pr]: https://github.com/github/github-store/compare
 [style]: https://github.com/styleguide/ruby
 [code-of-conduct]: CODE_OF_CONDUCT.md
