+++ Gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in github-kv.gemspec
+# Specify your gem's dependencies in github-store.gemspec
 gemspec
 
 gem "rails", "~> #{ENV['RAILS_VERSION'] || '5.0.2'}"
