+++ lib/generators/github/store/active_record_generator.rb
