+++ lib/generators/github/store/templates/migration.rb
