+++ lib/github-store.rb
@@ -1,0 +1,1 @@
+require "github/store"
