+++ lib/github/store.rb
@@ -1,0 +1,7 @@
+module GitHub
+  module Store
+  end
+end
+
+require "github/store/version"
+require "github/kv"
