+++ lib/github/store/version.rb
@@ -1,5 +1,5 @@
-module Github
-  class KV
+module GitHub
+  module Store
     VERSION = "0.1.0"
   end
 end
